Various changes
Added support for sections, since it was giving problems with filenames.
Removed youdontsurf from the pre-defined subreddit list, cause I
thought that was not something a lot of people liked.
Recognized that this didn’t work with galleries and made it so that it
just ignored galleries.
And plus added the subreddit name before the filename.