General Optimization (Details in Description)
KITTENS:
Cleaned up kittens a bit, added url retrieve, made it print the
resolutions so you know what you got out of it.

Pretty much a rewrite for kittens…

IMAGEPARSING:

Optimized archiving by having already archived images in a list, as not
to read the file for every image.

Because of that, added a “add_to_file” function.

Removed the whole “matches” argument from the main(), as it was really
of no use.

Moved the choice import to the top.
(Minor thing: Added “exception:” to the exception printing, which
shouldn’t print.)